requirements.txt
pytest==7.1.2
pytest-selenium
termcolor~=1.1.0
allure-python-commons
